Save sidebar files or folders on click outside
