+++ packages/web-ui/src/sections/Document/Sidebar/Files/NodeHeaderWrapper/index.tsx
@@ -103,7 +103,6 @@ function NodeHeaderWrapper({
   const { error, inputValue, onInputChange, onInputKeyDown } = useNodeValidator(
     {
       name,
-      nodeRef: nodeRef as RefObject<HTMLDivElement>,
       inputRef: inputRef as RefObject<HTMLInputElement>,
       isEditing,
       setIsEditing,
